plutus-book: skeleton
This adds a skeleton for the Plutus Book, with a spine file and
a demo chapter, as well as the Nix build and some instructions.

The Nix build creates the HTML, PDF, and EPUB version. We can
also do Kindle, but `kindlegen` has an unfree license, and I should
check that's okay.